♻️ move get_current_note to separate file
